requirements.txt
boto3==1.11.17
botocore==1.14.17
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.24
django-forms-bootstrap==3.1.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.4
Pillow==7.0.0
psycopg2==2.7.3.2
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
stripe==2.43.0
urllib3==1.25.8
whitenoise==5.0.1
